Fix: Changes to dragon Name as it now required and max string length is 15 with min 2. Dragons age changed to decimal from double so it only shows 1 place after decimal. GameSettings, GameStatusResponse, LifeCycleManagerShould amended to reflect the change. New migrations added.

+++ Tamagotchi/Contracts/GameStatusResponse.cs
@@ -8,7 +8,7 @@ public class GameStatusResponse
 
     public string DragonName { get; set; } = default!;
 
-    public double Age { get; set; }
+    public decimal Age { get; set; }
 
     public AgeGroup AgeGroup { get; set; }
